Revert "Feature: roadmap - private, description"
