+++ backend/roadmap/models.py
@@ -6,11 +6,9 @@ from user.models import User
 
 
 class Roadmap(models.Model):
-    private = models.BooleanField(default=False)
     title = models.CharField(max_length=64, default="")
     date = models.DateTimeField(auto_now_add=True)
     level = models.IntegerField(default=0)
-    description = models.TextField(default="")
     like_count = models.IntegerField(default=0)
     comment_count = models.IntegerField(default=0)
     pin_count = models.IntegerField(default=0)
