+++ frontend/.eslintrc.js
@@ -25,6 +25,5 @@ module.exports = {
     "jsx-a11y/label-has-associated-control": 0,
     "no-case-declarations": 0,
     "react/require-default-props": 0,
-    "react/jsx-wrap-multilines": 0,
   },
 };
