+++ frontend/package.json
@@ -4,7 +4,6 @@
   "private": true,
   "proxy": "http://localhost:8000/",
   "dependencies": {
-    "@material-ui/core": "^4.11.0",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
