+++ frontend/src/components/CreateSection/CreateSection.js
@@ -1,7 +1,6 @@
 import React from "react";
 import PropTypes from "prop-types";
 import CreateTask from "../CreateTask/CreateTask";
-import "./CreateSection.scss";
 
 const CreateSection = (props) => {
   const {
