+++ frontend/src/containers/CreateRoadmap/CreateRoadmap.js
@@ -4,17 +4,282 @@ import { withRouter } from "react-router";
 import PropTypes from "prop-types";
 import * as actionCreators from "../../store/actions/index";
 
-import Roadmap from "../Roadmap/Roadmap";
+import CreateSection from "../../components/CreateSection/CreateSection";
+import { levelType } from "../../constants";
+import "./CreateRoadmap.scss";
 
 class CreateRoadmap extends Component {
-  onClickCreateBack = () => {
-    const { history } = this.props;
-    history.goBack();
+  state = {
+    title: "",
+    level: 0,
+    sections: [],
+    tags: [],
+    newTag: "",
   };
 
-  onClickCreateConfirm = (roadmapData) => {
+  onChangeTitle = (title) => {
+    this.setState({ title });
+  };
+
+  onChangeLevel = (level) => {
+    this.setState({ level });
+  };
+
+  onChangeNewTag = (newTag) => {
+    this.setState({ newTag });
+  };
+
+  onClickAddTag = () => {
+    const { tags, newTag } = this.state;
+    this.setState({ tags: tags.concat(newTag) });
+    this.setState({ newTag: "" });
+  };
+
+  onClickDeleteTag = (id) => {
+    const { tags } = this.state;
+    this.setState({ tags: tags.filter((_, index) => index !== id) });
+  };
+
+  onClickCreateSection = () => {
+    const { sections } = this.state;
+    this.setState({ sections: sections.concat({ section_title: "", tasks: [] }) });
+  };
+
+  onClickDeleteSection = (tmpSectionId) => {
+    const { sections } = this.state;
+    this.setState({
+      sections: sections.filter((_, index) => {
+        return index !== tmpSectionId;
+      }),
+    });
+  };
+
+  onClickUpSection = (tmpSectionId) => {
+    const { sections } = this.state;
+    this.setState({
+      sections: sections.map((section, index) => {
+        if (index === tmpSectionId - 1) {
+          return sections[index + 1];
+        }
+        if (index === tmpSectionId) {
+          return sections[index - 1];
+        }
+        return section;
+      }),
+    });
+  };
+
+  onClickDownSection = (tmpSectionId) => {
+    const { sections } = this.state;
+    this.setState({
+      sections: sections.map((section, index) => {
+        if (index === tmpSectionId) {
+          return sections[index + 1];
+        }
+        if (index === tmpSectionId + 1) {
+          return sections[index - 1];
+        }
+        return section;
+      }),
+    });
+  };
+
+  onChangeSectionTitle = (tmpSectionId, title) => {
+    const { sections } = this.state;
+    this.setState({
+      sections: sections.map((section, index) => {
+        if (index === tmpSectionId) {
+          section.section_title = title;
+        }
+        return section;
+      }),
+    });
+  };
+
+  onClickCreateTask = (tmpSectionId) => {
+    const { sections } = this.state;
+    this.setState({
+      sections: sections.map((section, index) => {
+        if (index === tmpSectionId) {
+          return {
+            ...section,
+            tasks: section.tasks.concat({
+              task_title: "",
+              task_type: 0,
+              task_url: "",
+              task_description: "",
+            }),
+          };
+        }
+        return section;
+      }),
+    });
+  };
+
+  onClickDeleteTask = (tmpSectionId, tmpTaskId) => {
+    const { sections } = this.state;
+    this.setState({
+      sections: sections.map((section, sectionIndex) => {
+        if (sectionIndex === tmpSectionId) {
+          return {
+            ...section,
+            tasks: section.tasks.filter((_, taskIndex) => {
+              return taskIndex !== tmpTaskId;
+            }),
+          };
+        }
+        return section;
+      }),
+    });
+  };
+
+  onClickUpTask = (tmpSectionId, tmpTaskId) => {
+    const { sections } = this.state;
+    this.setState({
+      sections: sections.map((section, sectionIndex) => {
+        if (sectionIndex === tmpSectionId) {
+          return {
+            ...section,
+            tasks: section.tasks.map((task, taskIndex) => {
+              if (taskIndex === tmpTaskId - 1) {
+                return section.tasks[taskIndex + 1];
+              }
+              if (taskIndex === tmpTaskId) {
+                return section.tasks[taskIndex - 1];
+              }
+              return task;
+            }),
+          };
+        }
+        return section;
+      }),
+    });
+  };
+
+  onClickDownTask = (tmpSectionId, tmpTaskId) => {
+    const { sections } = this.state;
+    this.setState({
+      sections: sections.map((section, sectionIndex) => {
+        if (sectionIndex === tmpSectionId) {
+          return {
+            ...section,
+            tasks: section.tasks.map((task, taskIndex) => {
+              if (taskIndex === tmpTaskId) {
+                return section.tasks[taskIndex + 1];
+              }
+              if (taskIndex === tmpTaskId + 1) {
+                return section.tasks[taskIndex - 1];
+              }
+              return task;
+            }),
+          };
+        }
+        return section;
+      }),
+    });
+  };
+
+  onChangeTaskTitle = (tmpSectionId, tmpTaskId, title) => {
+    const { sections } = this.state;
+    this.setState({
+      sections: sections.map((section, sectionIndex) => {
+        if (sectionIndex === tmpSectionId) {
+          return {
+            ...section,
+            tasks: section.tasks.map((task, taskIndex) => {
+              if (taskIndex === tmpTaskId) {
+                return { ...task, task_title: title };
+              }
+              return task;
+            }),
+          };
+        }
+        return section;
+      }),
+    });
+  };
+
+  onChangeTaskType = (tmpSectionId, tmpTaskId, type) => {
+    const { sections } = this.state;
+    this.setState({
+      sections: sections.map((section, sectionIndex) => {
+        if (sectionIndex === tmpSectionId) {
+          return {
+            ...section,
+            tasks: section.tasks.map((task, taskIndex) => {
+              if (taskIndex === tmpTaskId) {
+                return { ...task, task_type: type };
+              }
+              return task;
+            }),
+          };
+        }
+        return section;
+      }),
+    });
+  };
+
+  onChangeTaskUrl = (tmpSectionId, tmpTaskId, url) => {
+    const { sections } = this.state;
+    this.setState({
+      sections: sections.map((section, sectionIndex) => {
+        if (sectionIndex === tmpSectionId) {
+          return {
+            ...section,
+            tasks: section.tasks.map((task, taskIndex) => {
+              if (taskIndex === tmpTaskId) {
+                return { ...task, task_url: url };
+              }
+              return task;
+            }),
+          };
+        }
+        return section;
+      }),
+    });
+  };
+
+  onChangeTaskDescription = (tmpSectionId, tmpTaskId, description) => {
+    const { sections } = this.state;
+    this.setState({
+      sections: sections.map((section, sectionIndex) => {
+        if (sectionIndex === tmpSectionId) {
+          return {
+            ...section,
+            tasks: section.tasks.map((task, taskIndex) => {
+              if (taskIndex === tmpTaskId) {
+                return { ...task, task_description: description };
+              }
+              return task;
+            }),
+          };
+        }
+        return section;
+      }),
+    });
+  };
+
+  onClickCreateBack = () => {
+    const { history } = this.props;
+    const { title, level, sections } = this.state;
+    if (title !== "" || level !== 0 || sections.length !== 0) {
+      const back = window.confirm("Leave the page? Changes you made will be deleted.");
+      if (back) {
+        history.goBack();
+      }
+    } else history.goBack();
+  };
+
+  onClickCreateConfirm = () => {
+    const { title, level, sections, tags } = this.state;
     const { onCreateRoadmap } = this.props;
-    onCreateRoadmap({ ...roadmapData, addedTagList: undefined, deletedTagList: undefined });
+    const roadmapData = {
+      title,
+      level,
+      sections,
+      tags,
+    };
+    onCreateRoadmap(roadmapData);
   };
 
   render() {
@@ -25,28 +290,114 @@ class CreateRoadmap extends Component {
       return <div />;
     }
 
+    const { sections, level, title, tags, newTag } = this.state;
+
+    const taglist = tags.map((tag, index) => {
+      return (
+        <div className="tags">
+          <p key={tag}>{tag}</p>
+          <button className="delete-tag" type="button" onClick={() => this.onClickDeleteTag(index)}>
+            delete
+          </button>
+        </div>
+      );
+    });
+
+    const CreateSections = sections.map((section, index) => {
+      return (
+        <CreateSection
+          tmpSectionId={index}
+          sectionLastId={sections.length - 1}
+          title={section.section_title}
+          tasks={section.tasks}
+          clickDeleteSectionHandler={this.onClickDeleteSection}
+          clickUpSectionHandler={this.onClickUpSection}
+          clickDownSectionHandler={this.onClickDownSection}
+          changeSectionTitleHandler={this.onChangeSectionTitle}
+          clickCreateTaskHandler={this.onClickCreateTask}
+          clickDeleteTaskHandler={this.onClickDeleteTask}
+          clickUpTaskHandler={this.onClickUpTask}
+          clickDownTaskHandler={this.onClickDownTask}
+          changeTaskTitleHandler={this.onChangeTaskTitle}
+          changeTaskTypeHandler={this.onChangeTaskType}
+          changeTaskUrlHandler={this.onChangeTaskUrl}
+          changeTaskDescriptionHandler={this.onChangeTaskDescription}
+        />
+      );
+    });
+
     return (
       <div className="CreateRoadmap">
-        <Roadmap
-          isEdit={false}
-          onClickBackHandler={this.onClickCreateBack}
-          onClickConfirmHandler={this.onClickCreateConfirm}
-        />
+        <h1 className="header">Create Roadmap</h1>
+        <div className="roadmap">
+          <label>Title</label>
+          <input
+            id="roadmap-title"
+            type="text"
+            value={title}
+            onChange={(event) => this.onChangeTitle(event.target.value)}
+          />
+          <select
+            id="roadmap-level"
+            value={level}
+            onChange={(event) => this.onChangeLevel(event.target.value)}
+          >
+            Level
+            <option value={0}>Choose level</option>
+            <option value={levelType.BASIC}>Basic</option>
+            <option value={levelType.INTERMEDIATE}>Intermediate</option>
+            <option value={levelType.ADVANCED}>Advanced</option>
+          </select>
+          <br />
+          <label>Tags</label>
+          {taglist}
+          <input
+            id="new-tag"
+            value={newTag}
+            onChange={(event) => this.onChangeNewTag(event.target.value)}
+          />
+          <button id="add-tag-button" type="button" onClick={() => this.onClickAddTag()}>
+            add
+          </button>
+        </div>
+        <div className="sections">
+          {CreateSections}
+          <button
+            type="button"
+            id="create-section-button"
+            onClick={() => {
+              this.onClickCreateSection();
+            }}
+          >
+            Create Section
+          </button>
+        </div>
+        <div className="buttons">
+          <button
+            id="back-create-roadmap-button"
+            type="button"
+            onClick={() => this.onClickCreateBack()}
+          >
+            Back
+          </button>
+          <button
+            id="confirm-create-roadmap-button"
+            type="button"
+            disabled={title === "" || level === 0 || sections.length === 0}
+            onClick={() => this.onClickCreateConfirm()}
+          >
+            Confirm
+          </button>
+        </div>
       </div>
     );
   }
 }
 
 CreateRoadmap.propTypes = {
+  history: PropTypes.objectOf(PropTypes.any),
   selectedUser: PropTypes.objectOf(PropTypes.any),
   onCreateRoadmap: PropTypes.func.isRequired,
-  history: PropTypes.objectOf(PropTypes.any),
-};
-
-const mapStateToProps = (state) => {
-  return {
-    selectedUser: state.user.selectedUser,
-  };
 };
 
 const mapDispatchToProps = (dispatch) => {
@@ -55,4 +406,4 @@ const mapDispatchToProps = (dispatch) => {
   };
 };
 
-export default connect(mapStateToProps, mapDispatchToProps)(withRouter(CreateRoadmap));
+export default connect(null, mapDispatchToProps)(withRouter(CreateRoadmap));
