+++ frontend/src/containers/CreateRoadmap/CreateRoadmap.test.js
@@ -7,7 +7,6 @@ import { ConnectedRouter } from "connected-react-router";
 import getMockStore from "../../test-utils/mocks";
 import { history } from "../../store/store";
 import CreateRoadmap from "./CreateRoadmap";
-import Roadmap from "../Roadmap/Roadmap";
 import CreateSection from "../../components/CreateSection/CreateSection";
 import * as actionCreators from "../../store/actions/roadmap";
 
@@ -81,8 +80,6 @@ describe("<CreateRoadmap />", () => {
     const title = "test-title";
     const level = 2;
     const newTag = "test";
-    const isPrivate = true;
-    const description = "test";
     const component = mount(createRoadmap);
     let wrapper = component.find("#roadmap-title");
     wrapper.simulate("change", { target: { value: title } });
@@ -90,16 +87,10 @@ describe("<CreateRoadmap />", () => {
     wrapper.simulate("change", { target: { value: level } });
     wrapper = component.find("#new-tag");
     wrapper.simulate("change", { target: { value: newTag } });
-    wrapper = component.find("#roadmap-private");
-    wrapper.at(2).simulate("click");
-    wrapper = component.find("#roadmap-description");
-    wrapper.simulate("change", { target: { value: description } });
-    const instance = component.find(Roadmap).instance();
+    const instance = component.find(CreateRoadmap.WrappedComponent).instance();
     expect(instance.state.title).toBe(title);
     expect(instance.state.level).toBe(level);
     expect(instance.state.newTag).toBe(newTag);
-    expect(instance.state.isPrivate).toBe(isPrivate);
-    expect(instance.state.description).toBe(description);
   });
 
   it("should call 'onClickAddTag', 'onClickDeleteTag'", () => {
@@ -109,7 +100,7 @@ describe("<CreateRoadmap />", () => {
     wrapper.simulate("change", { target: { value: newTag } });
     wrapper = component.find("#add-tag-button");
     wrapper.simulate("click");
-    const instance = component.find(Roadmap).instance();
+    const instance = component.find(CreateRoadmap.WrappedComponent).instance();
     expect(instance.state.tags).toEqual([newTag]);
     wrapper = component.find(".delete-tag");
     wrapper.at(0).simulate("click");
@@ -120,7 +111,7 @@ describe("<CreateRoadmap />", () => {
     const component = mount(createRoadmap);
     let wrapper = component.find("#create-section-button");
     wrapper.simulate("click");
-    const instance = component.find(Roadmap).instance();
+    const instance = component.find(CreateRoadmap.WrappedComponent).instance();
     expect(instance.state.sections[0]).toEqual({ section_title: "", tasks: [] });
     wrapper = component.find(CreateSection);
     expect(wrapper.length).toBe(1);
@@ -132,7 +123,7 @@ describe("<CreateRoadmap />", () => {
     const component = mount(createRoadmap);
     let wrapper = component.find("#create-section-button");
     wrapper.simulate("click");
-    const instance = component.find(Roadmap).instance();
+    const instance = component.find(CreateRoadmap.WrappedComponent).instance();
     const sectionsNum = instance.state.sections.length;
     wrapper = component.find(".delete-section-button");
     wrapper.simulate("click");
@@ -145,7 +136,7 @@ describe("<CreateRoadmap />", () => {
     wrapper.simulate("click");
     wrapper = component.find(".section-title");
     wrapper.simulate("change", { target: { value: "1" } });
-    const instance = component.find(Roadmap).instance();
+    const instance = component.find(CreateRoadmap.WrappedComponent).instance();
     expect(instance.state.sections[0].section_title).toBe("1");
   });
 
@@ -161,7 +152,7 @@ describe("<CreateRoadmap />", () => {
     wrapper.at(2).simulate("change", { target: { value: "2" } });
     wrapper = component.find(".up-section-button");
     wrapper.at(1).simulate("click");
-    const instance = component.find(Roadmap).instance();
+    const instance = component.find(CreateRoadmap.WrappedComponent).instance();
     expect(instance.state.sections[0].section_title).toBe("1");
     expect(instance.state.sections[1].section_title).toBe("0");
     expect(instance.state.sections[2].section_title).toBe("2");
@@ -179,7 +170,7 @@ describe("<CreateRoadmap />", () => {
     wrapper.at(2).simulate("change", { target: { value: "2" } });
     wrapper = component.find(".down-section-button");
     wrapper.at(0).simulate("click");
-    const instance = component.find(Roadmap).instance();
+    const instance = component.find(CreateRoadmap.WrappedComponent).instance();
     expect(instance.state.sections[0].section_title).toBe("1");
     expect(instance.state.sections[1].section_title).toBe("0");
     expect(instance.state.sections[2].section_title).toBe("2");
@@ -192,7 +183,7 @@ describe("<CreateRoadmap />", () => {
     wrapper.simulate("click");
     wrapper = component.find(".create-task-button");
     wrapper.at(0).simulate("click");
-    const instance = component.find(Roadmap).instance();
+    const instance = component.find(CreateRoadmap.WrappedComponent).instance();
     expect(instance.state.sections[0].tasks[0]).toEqual({
       task_title: "",
       task_type: 0,
@@ -210,7 +201,7 @@ describe("<CreateRoadmap />", () => {
     wrapper = component.find(".create-task-button");
     wrapper.at(0).simulate("click");
     wrapper.at(1).simulate("click");
-    const instance = component.find(Roadmap).instance();
+    const instance = component.find(CreateRoadmap.WrappedComponent).instance();
     const tasksNum0 = instance.state.sections[0].tasks.length;
     const tasksNum1 = instance.state.sections[1].tasks.length;
     wrapper = component.find(".delete-task-button");
@@ -230,7 +221,7 @@ describe("<CreateRoadmap />", () => {
     wrapper.at(0).simulate("click");
     wrapper = component.find(".task-title");
     wrapper.at(0).simulate("change", { target: { value: testTitle } });
-    const instance = component.find(Roadmap).instance();
+    const instance = component.find(CreateRoadmap.WrappedComponent).instance();
     expect(instance.state.sections[0].tasks[0].task_title).toEqual(testTitle);
     expect(instance.state.sections[0].tasks[1].task_title).toEqual("");
     expect(instance.state.sections[1]).toEqual({ section_title: "", tasks: [] });
@@ -247,7 +238,7 @@ describe("<CreateRoadmap />", () => {
     wrapper.at(0).simulate("click");
     wrapper = component.find(".task-type");
     wrapper.at(0).simulate("change", { target: { value: testType } });
-    const instance = component.find(Roadmap).instance();
+    const instance = component.find(CreateRoadmap.WrappedComponent).instance();
     expect(instance.state.sections[0].tasks[0].task_type).toEqual(testType);
     expect(instance.state.sections[0].tasks[1].task_type).toEqual(0);
     expect(instance.state.sections[1]).toEqual({ section_title: "", tasks: [] });
@@ -264,7 +255,7 @@ describe("<CreateRoadmap />", () => {
     wrapper.at(0).simulate("click");
     wrapper = component.find(".task-url");
     wrapper.at(0).simulate("change", { target: { value: testUrl } });
-    const instance = component.find(Roadmap).instance();
+    const instance = component.find(CreateRoadmap.WrappedComponent).instance();
     expect(instance.state.sections[0].tasks[0].task_url).toEqual(testUrl);
     expect(instance.state.sections[0].tasks[1].task_url).toEqual("");
     expect(instance.state.sections[1]).toEqual({ section_title: "", tasks: [] });
@@ -281,7 +272,7 @@ describe("<CreateRoadmap />", () => {
     wrapper.at(0).simulate("click");
     wrapper = component.find(".task-description");
     wrapper.at(0).simulate("change", { target: { value: testDescription } });
-    const instance = component.find(Roadmap).instance();
+    const instance = component.find(CreateRoadmap.WrappedComponent).instance();
     expect(instance.state.sections[0].tasks[0].task_description).toEqual(testDescription);
     expect(instance.state.sections[0].tasks[1].task_description).toEqual("");
     expect(instance.state.sections[1]).toEqual({ section_title: "", tasks: [] });
@@ -289,7 +280,7 @@ describe("<CreateRoadmap />", () => {
 
   it("should call 'onClickUpTask'", () => {
     const component = mount(createRoadmap);
-    const instance = component.find(Roadmap).instance();
+    const instance = component.find(CreateRoadmap.WrappedComponent).instance();
     let wrapper = component.find("#create-section-button");
     wrapper.simulate("click");
     wrapper.simulate("click");
@@ -311,7 +302,7 @@ describe("<CreateRoadmap />", () => {
 
   it("should call 'onClickDownTask'", () => {
     const component = mount(createRoadmap);
-    const instance = component.find(Roadmap).instance();
+    const instance = component.find(CreateRoadmap.WrappedComponent).instance();
     let wrapper = component.find("#create-section-button");
     wrapper.simulate("click");
     wrapper.simulate("click");
@@ -333,7 +324,7 @@ describe("<CreateRoadmap />", () => {
 
   it("should call 'onClickCreateBack' - no confirm", () => {
     const component = mount(createRoadmap);
-    const wrapper = component.find("#back-roadmap-button");
+    const wrapper = component.find("#back-create-roadmap-button");
     wrapper.simulate("click");
     expect(spyHistoryGoBack).toHaveBeenCalledTimes(1);
     spyHistoryGoBack.mockRestore();
@@ -342,7 +333,7 @@ describe("<CreateRoadmap />", () => {
   it("should call 'onClickCreateBack' - confirm=True", () => {
     const spyBackConfirmTrue = jest.spyOn(window, "confirm").mockImplementation(() => true);
     const component = mount(createRoadmap);
-    const wrapper = component.find("#back-roadmap-button");
+    const wrapper = component.find("#back-create-roadmap-button");
     const wrapper2 = component.find("#create-section-button");
     wrapper2.simulate("click");
     wrapper.simulate("click");
@@ -354,7 +345,7 @@ describe("<CreateRoadmap />", () => {
   it("should call 'onClickCreateBack' - confirm=False", () => {
     const spyBackConfirmFalse = jest.spyOn(window, "confirm").mockImplementation(() => false);
     const component = mount(createRoadmap);
-    const wrapper = component.find("#back-roadmap-button");
+    const wrapper = component.find("#back-create-roadmap-button");
     const wrapper2 = component.find("#create-section-button");
     wrapper2.simulate("click");
     wrapper.simulate("click");
@@ -374,7 +365,7 @@ describe("<CreateRoadmap />", () => {
     wrapper.simulate("change", { target: { value: testLevel } });
     wrapper = component.find("#create-section-button");
     wrapper.simulate("click");
-    wrapper = component.find("#confirm-roadmap-button");
+    wrapper = component.find("#confirm-create-roadmap-button");
     wrapper.simulate("click");
     expect(spyCreateRoadmap).toHaveBeenCalledTimes(1);
   });
